move gql out to file
