+++ frontend/src/components/Dashboard/CrossingDetailPage/CrossingDetailPage.js
@@ -3,7 +3,8 @@ import { graphql, compose } from 'react-apollo';
 import gql from 'graphql-tag';
 import CrossingStaticMap from 'components/Map/CrossingStaticMap';
 import CrossingDetails from 'components/Dashboard/CrossingDetailPage/CrossingDetails';
-import CrossingStatusHistory from 'components/Dashboard/CrossingDetailPage/CrossingStatusHistory';
+import CrossingStatusHistory from 'components/Dashboard/CrossingStatusHistory/CrossingStatusHistory';
+import statusHistoryQuery from 'components/Dashboard/CrossingListPage/queries/statusHistoryQuery';
 import { ContainerQuery } from 'react-container-query';
 import classnames from 'classnames';
 import { LARGE_ITEM_MIN_WIDTH } from 'constants/containerQueryConstants';
@@ -20,15 +21,15 @@ class CrossingDetailPage extends Component {
     const isLoading = (
       !this.props.CrossingByIdQuery ||
        this.props.CrossingByIdQuery.loading ||
-      !this.props.CrossingHistoryQuery ||
-       this.props.CrossingHistoryQuery.loading
+      !this.props.StatusHistoryQuery ||
+       this.props.StatusHistoryQuery.loading
     );
 
     if ( isLoading ) { return (<div>Loading</div>) };
 
     const crossing = this.props.CrossingByIdQuery.crossingById;
     const communities = crossing.communityCrossingsByCrossingId.nodes.map(n => n.communityByCommunityId);
-    const history = this.props.CrossingHistoryQuery.allStatusUpdates.nodes;
+    const history = this.props.StatusHistoryQuery.allStatusUpdates.nodes;
 
     return (
       <ContainerQuery query={containerQuery}>
@@ -72,35 +73,6 @@ const CrossingByIdQuery = gql`
   }
 `;
 
-const CrossingHistoryQuery = gql`
-query crossingHistory($crossingId:Int!) {
-  allStatusUpdates(
-    condition: { crossingId: $crossingId }
-    orderBy: CREATED_AT_DESC
-  ) {
-    nodes {
-      userByCreatorId {
-        id
-        lastName
-        firstName
-      }
-      statusByStatusId {
-        id
-        name
-      }
-      statusReasonByStatusReasonId {
-        name
-      }
-      statusDurationByStatusDurationId {
-         name
-       }
-      createdAt
-      notes
-    }
-  }
-}
-`;
-
 export default compose(
   graphql(CrossingByIdQuery, {
     name: 'CrossingByIdQuery',
@@ -110,8 +82,8 @@ export default compose(
       }
     })
   }),
-  graphql(CrossingHistoryQuery, {
-    name: 'CrossingHistoryQuery',
+  graphql(statusHistoryQuery, {
+    name: 'StatusHistoryQuery',
     options: (ownProps) => ({
       variables: {
         crossingId: ownProps.match.params.id
