+++ frontend/src/components/Dashboard/CrossingStatusHistory/CrossingStatusHistory.scss
