+++ frontend/src/components/Dashboard/CrossingStatusHistoryPage/CrossingStatusHistoryPage.js
@@ -3,7 +3,8 @@ import { graphql, compose } from 'react-apollo';
 import gql from 'graphql-tag';
 import CrossingStaticMap from 'components/Map/CrossingStaticMap';
 import CrossingDetails from 'components/Dashboard/CrossingDetailPage/CrossingDetails';
-import CrossingStatusHistory from 'components/Dashboard/CrossingDetailPage/CrossingStatusHistory';
+import CrossingStatusHistory from 'components/Dashboard/CrossingStatusHistory/CrossingStatusHistory';
+import statusHistoryQuery from 'components/Dashboard/CrossingListPage/queries/statusHistoryQuery';
 import { ContainerQuery } from 'react-container-query';
 import classnames from 'classnames';
 import { LARGE_ITEM_MIN_WIDTH } from 'constants/containerQueryConstants';
@@ -30,46 +31,16 @@ class CrossingStatusHistoryPage extends Component {
       <ContainerQuery query={containerQuery}>
         {(params) => (
           <div className="CrossingDetailPage">
-            <CrossingStatusHistory history={history}/>
+            <CrossingStatusHistory history={history} showNames/>
           </div>
         )}
       </ContainerQuery>
     );
   }
-
 }
 
-const CrossingHistoryQuery = gql`
-query crossingHistory {
-  allStatusUpdates(
-    orderBy: CREATED_AT_DESC
-  ) {
-    nodes {
-      userByCreatorId {
-        id
-        lastName
-        firstName
-      }
-      statusByStatusId {
-        id
-        name
-      }
-      statusReasonByStatusReasonId {
-        name
-      }
-      statusDurationByStatusDurationId {
-         name
-       }
-      createdAt
-      notes
-      crossingId
-      crossingByCrossingId {
-        name
-        humanAddress
-      }
-    }
-  }
-}
-`;
-
-export default graphql(CrossingHistoryQuery)(CrossingStatusHistoryPage);
+export default graphql(statusHistoryQuery, {
+  options: (ownProps) => ({
+    variables: {}
+  })
+})(CrossingStatusHistoryPage);
